Upgrading to latest Angular 2 and CLI versions
Updating the project structure to the latest Angular CLI 1.0.0.beta-1
version and to the latest release candidate of Angular 2.

As part of this update all of the Angular 2 imports have changed and
I updated the code to use the non-deprecated router by looking through
its source in the Angular 2 repository on GitHub.